Diabetes_DS/requirements.txt
streamlit
pandas
scikit-learn
xgboost
shap
matplotlib
joblib
Pillow  # for SHAP image display via PIL
